backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "celery[redis]>=5.6.3",
    "fastapi>=0.136.0",
    "geoalchemy2>=0.19.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "ortools>=9.12.4544",
    "passlib>=1.7.4",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic[email]>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-ulid>=3.1.0",
    "redis[hiredis]>=6.4.0",
    "spacy>=3.8.14",
    "sqlalchemy[asyncio]>=2.0.49",
    "razorpay>=1.4.1",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.45.0",
    "pybreaker>=1.4.1",
    "feedparser>=6.0.12",
    "firebase-admin>=6.0.0",
    "numpy>=2.4.4",
    "langgraph>=1.1.9",
    "langchain-google-genai>=4.2.2",
    "langchain-core>=1.3.2",
    "google-genai>=1.73.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "coverage>=7.13.5",
    "fakeredis>=2.35.1",
    "httpx[asyncio]>=0.28.1",
    "mypy>=1.20.2",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "S", "B", "A", "C4", "T20"]
ignore = [
    "B008",   # FastAPI Depends() in default args — intentional pattern
    "S311",   # random module — used for simulation/seeding only, not cryptography
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S311", "S608", "S105", "S106", "T201", "E501"]
"db/migrations/*" = ["E501", "I001", "F401", "S608"]
"_fix_models.py" = ["T201", "E401", "I001", "F401"]
"db/seed.py" = ["T201", "S101"]
"init_db.py" = ["T201"]
"scripts/*" = ["T201", "E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = ["ignore::DeprecationWarning"]


[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*", "db/migrations/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.mypy]
strict = true
ignore_missing_imports = true
